requirements.txt
selenium==4.25.0
pytest==8.3.3
pytest-html==4.1.1
webdriver-manager==4.0.2
pyyaml==6.0.2